Voting refactored
